// src/api/models/IHistoricoContactoRRHH.ts
export interface IHistoricoContactoRRHH {
  id?: number;
  id_contacto: number;
  usuario: string;
  fecha: string;
  notas?: string;
  ultima_llamada?: string;
  siguiente_paso?: string;
  nombre?: string;
  apellido?: string;
  telefono?: string;
  telefono2?: string;
  universidad?: string;
  tipo?: string;
  puesto?: string;
  email?: string;
  portal_web?: string;
  departamento?: string;
  usuario_portal?: string;
  contrasena_portal?: string;
  firma_convenio_fecha?: string;
  firma_convenio_link?: string;
  vencimiento_convenio?: string;
  altas_social?: string;
}